Refactoring likeihood computation for subst models
Refactoring likelihood computation, moved all relevant functions to
the substitution model module.
