+++ phylo/src/substitution_models/mod.rs
@@ -1,8 +1,14 @@
-use crate::{f64_h, Result};
-use anyhow::anyhow;
-use nalgebra::{Const, DimMin, SMatrix, SVector};
+use crate::{
+    f64_h,
+    likelihood::{EvolutionaryModelInfo, LikelihoodCostFunction},
+    phylo_info::PhyloInfo,
+    sequences::NUCLEOTIDES_STR,
+    tree::NodeIdx,
+};
+use nalgebra::{Const, DMatrix, DimMin, SMatrix, SVector};
 use ordered_float::OrderedFloat;
 use std::collections::HashMap;
+use std::ops::Mul;
 
 pub mod dna_models;
 pub mod protein_models;
@@ -15,47 +21,6 @@ pub struct SubstitutionModel<const N: usize> {
     index: [i32; 255],
     pub q: SubstMatrix<N>,
     pub pi: FreqVector<N>,
-}
-
-pub type DNASubstModel = SubstitutionModel<4>;
-pub type ProteinSubstModel = SubstitutionModel<20>;
-
-impl DNASubstModel {
-    pub fn new(model_name: &str, model_params: &[f64]) -> Result<Self> {
-        let q: SubstMatrix<4>;
-        let pi: FreqVector<4>;
-        match model_name.to_uppercase().as_str() {
-            "JC69" => (q, pi) = dna_models::jc69(model_params)?,
-            "K80" => (q, pi) = dna_models::k80(model_params)?,
-            "TN93" => (q, pi) = dna_models::tn93(model_params)?,
-            "GTR" => (q, pi) = dna_models::gtr(model_params)?,
-            _ => return Err(anyhow!("Unknown DNA model requested.")),
-        }
-        let model = DNASubstModel {
-            index: dna_models::nucleotide_index(),
-            q,
-            pi,
-        };
-        Ok(model)
-    }
-}
-
-impl ProteinSubstModel {
-    pub fn new(model_name: &str) -> Result<Self> {
-        let q: SubstMatrix<20>;
-        let pi: FreqVector<20>;
-        match model_name.to_uppercase().as_str() {
-            "WAG" => (q, pi) = protein_models::wag()?,
-            "BLOSUM" => (q, pi) = protein_models::blosum()?,
-            "HIVB" => (q, pi) = protein_models::hivb()?,
-            _ => return Err(anyhow!("Unknown protein model requested.")),
-        }
-        Ok(ProteinSubstModel {
-            index: protein_models::aminoacid_index(),
-            q,
-            pi,
-        })
-    }
 }
 
 impl<const N: usize> SubstitutionModel<N>
@@ -117,6 +82,132 @@ where
             scores.fill_diagonal(0.0);
         }
         (scores, scores.mean())
+    }
+}
+
+pub(crate) struct SubstitutionLikelihoodCost<'a, const N: usize> {
+    pub(crate) info: &'a PhyloInfo,
+    pub(crate) model: SubstitutionModel<N>,
+    pub(crate) temp_values: SubstitutionModelInfo<N>,
+}
+
+pub(crate) struct SubstitutionModelInfo<const N: usize> {
+    internal_info: Vec<DMatrix<f64>>,
+    internal_info_valid: Vec<bool>,
+    internal_models: Vec<SubstMatrix<N>>,
+    internal_models_valid: Vec<bool>,
+    leaf_info: Vec<DMatrix<f64>>,
+    leaf_info_valid: Vec<bool>,
+    leaf_models: Vec<SubstMatrix<N>>,
+    leaf_models_valid: Vec<bool>,
+    leaf_sequence_info: Vec<DMatrix<f64>>,
+}
+
+// TODO: Convert SubstitutionModel to a trait EvolutionaryModel, and implement it for DNASubstModel, ProteinSubstModel, and PIP
+// implies that the sequences are aligned
+impl<const N: usize> EvolutionaryModelInfo<N> for SubstitutionModelInfo<N> {
+    fn new(info: &PhyloInfo, model: &SubstitutionModel<N>) -> Self {
+        let leaf_count = info.tree.leaves.len();
+        let internal_count = info.tree.internals.len();
+        let msa_length = info.sequences[0].seq().len();
+        // set up basic char probabilities for the leaves
+        let leaf_sequence_info = info
+            .sequences
+            .iter()
+            .map(|rec| {
+                // This is incorrect (ignoress ambig chars) and only works for DNA
+                DMatrix::from_fn(4, msa_length, |i, j| match rec.seq()[j] {
+                    b'-' => model.pi[i],
+                    _ => {
+                        if NUCLEOTIDES_STR.find(rec.seq()[j] as char).unwrap() == i {
+                            1.0
+                        } else {
+                            0.0
+                        }
+                    }
+                })
+            })
+            .collect::<Vec<_>>();
+        SubstitutionModelInfo {
+            internal_info: vec![DMatrix::<f64>::zeros(N, msa_length); internal_count],
+            internal_info_valid: vec![false; internal_count],
+            internal_models: vec![SubstMatrix::zeros(); internal_count],
+            internal_models_valid: vec![false; internal_count],
+            leaf_info: vec![DMatrix::<f64>::zeros(N, msa_length); leaf_count],
+            leaf_info_valid: vec![false; leaf_count],
+            leaf_models: vec![SubstMatrix::zeros(); leaf_count],
+            leaf_models_valid: vec![false; leaf_count],
+            leaf_sequence_info,
+        }
+    }
+}
+
+impl<const N: usize> LikelihoodCostFunction<N> for SubstitutionLikelihoodCost<'_, N>
+where
+    Const<N>: DimMin<Const<N>, Output = Const<N>>,
+{
+    fn compute_log_likelihood(&mut self) -> f64 {
+        for node_idx in &self.info.tree.postorder {
+            match node_idx {
+                NodeIdx::Internal(idx) => {
+                    if !self.temp_values.internal_info_valid[*idx] {
+                        self.set_internal_values(idx);
+                    }
+                }
+                NodeIdx::Leaf(idx) => {
+                    if !self.temp_values.leaf_info_valid[*idx] {
+                        self.set_child_values(idx);
+                    }
+                }
+            };
+        }
+        let root_info = match self.info.tree.root {
+            NodeIdx::Internal(idx) => &self.temp_values.internal_info[idx],
+            NodeIdx::Leaf(idx) => &self.temp_values.leaf_info[idx],
+        };
+        let likelihood = self.model.pi.transpose().mul(root_info);
+        assert_eq!(likelihood.ncols(), 1);
+        assert_eq!(likelihood.nrows(), 1);
+        likelihood[(0, 0)].ln()
+    }
+}
+
+impl<'a, const N: usize> SubstitutionLikelihoodCost<'a, N>
+where
+    Const<N>: DimMin<Const<N>, Output = Const<N>>,
+{
+    fn set_internal_values(&mut self, idx: &usize) {
+        let node = &self.info.tree.internals[*idx];
+        if !self.temp_values.internal_models_valid[*idx] {
+            self.temp_values.internal_models[*idx] = self.model.get_p(node.blen);
+            self.temp_values.internal_models_valid[*idx] = true;
+        }
+        let childx_info = self.child_info(&node.children[0]);
+        let childy_info = self.child_info(&node.children[1]);
+        self.temp_values.internal_models[*idx].mul_to(
+            &(childx_info.component_mul(childy_info)),
+            &mut self.temp_values.internal_info[*idx],
+        );
+        self.temp_values.internal_info_valid[*idx] = true;
+    }
+
+    fn set_child_values(&mut self, idx: &usize) {
+        if !self.temp_values.leaf_models_valid[*idx] {
+            self.temp_values.leaf_models[*idx] = self.model.get_p(self.info.tree.leaves[*idx].blen);
+            self.temp_values.leaf_models_valid[*idx] = true;
+        }
+        self.temp_values.leaf_models[*idx].mul_to(
+            &self.temp_values.leaf_sequence_info[*idx],
+            &mut self.temp_values.leaf_info[*idx],
+        );
+        self.temp_values.leaf_info_valid[*idx] = true;
+    }
+
+    fn child_info(&self, child: &NodeIdx) -> &DMatrix<f64> {
+        match child {
+            NodeIdx::Internal(idx) => &self.temp_values.internal_info[*idx],
+            NodeIdx::Leaf(idx) => &self.temp_values.leaf_info[*idx],
+        }
     }
 }
